Move ready event to own file
